requirement.txt
accelerate
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
altair
annotated-types
antlr4-python3-runtime
anyio
attrs
backoff
bcrypt
beautifulsoup4
blinker
build
cachetools
certifi
cffi
chardet
charset-normalizer
chromadb
click
colorama
coloredlogs
contourpy
cryptography
cycler
dataclasses-json
Deprecated
distro
docx-charts
docx2txt
durationpy
effdet
emoji
et_xmlfile
fastapi
filelock
filetype
flatbuffers
fonttools
frozenlist
fsspec
gitdb
GitPython
google-api-core
google-auth
google-cloud-vision
googleapis-common-protos
greenlet
groq
grpcio
grpcio-status
gunicorn
h11
html5lib
httpcore
httptools
httpx
httpx-sse
huggingface-hub
humanfriendly
idna
importlib_metadata
importlib_resources
Jinja2
jiter
joblib
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
kiwisolver
kubernetes
langchain
langchain-chroma
langchain-community
langchain-core
langchain-experimental
langchain-groq
langchain-huggingface
langchain-ollama
langchain-openai
langchain-text-splitters
langdetect
langsmith
lxml
markdown-it-py
MarkupSafe
marshmallow
matplotlib
mdurl
mmh3
mpmath
multidict
mypy_extensions
narwhals
nest-asyncio
networkx
nltk
numpy
oauthlib
olefile
ollama
omegaconf
onnx
onnxruntime
openai
opencv-python
openpyxl
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
orjson
overrides
packaging
pandas
pdf2image
pdfminer.six
pi_heif
pikepdf
pillow
posthog
propcache
proto-plus
protobuf
psutil
pyarrow
pyasn1
pyasn1_modules
pybase64
pycocotools
pycparser
pydantic
pydantic-settings
pydantic_core
pydeck
Pygments
PyMuPDF
pyparsing
pypdf
pypdfium2
PyPika
pyproject_hooks
pyreadline3
pytesseract
python-dateutil
python-docx
python-dotenv
python-iso639
python-magic
python-multipart
python-oxmsg
python-pptx
pytz
PyYAML
RapidFuzz
referencing
regex
requests
requests-oauthlib
requests-toolbelt
rich
rpds-py
rsa
safetensors
scikit-learn
scipy
sentence-transformers
setuptools
shellingham
six
smmap
sniffio
soupsieve
SQLAlchemy
starlette
streamlit
sympy
tenacity
threadpoolctl
tiktoken
timm
tokenizers
toml
tools
torch
torchvision
tornado
tqdm
transformers
typer
typing-inspect
typing-inspection
typing_extensions
tzdata
unstructured-client
unstructured-inference
unstructured.pytesseract
urllib3
uvicorn
watchdog
watchfiles
webencodings
websocket-client
websockets
wrapt
xlsxwriter
yarl
zipp
zstandard
